start page after login + Int to Long

+++ src/main/java/com/vh/activitymanage/service/UserService.java
@@ -7,5 +7,5 @@ public interface UserService {
 
     User findByUsername(String username);
 
-    void deleteById(Integer id);
+    void deleteById(Long id);
 }
